Made large refactoring
