+++ Algorithms-and-Data-Structures/PR6-DataSortingAlgorithmImplementation/abstract_classes.py
@@ -1,8 +1,7 @@
 from abc import ABC, abstractmethod
 
 
-class ListAbsCommands(ABC):
-
+class AbstractList(ABC):
     @abstractmethod
     def add(self, value: object, index: int) -> bool:
         ...
